chore: update dependencies and improve validation tests
- Added `gmana_lints` as a dev dependency in `gmana_validation` and `gmana_value_objects`.
- Updated `analysis_options.yaml` to include `gmana_lints`.
- Refactored email, number, password, and money validators to use `const` constructors where applicable.
- Improved test cases for email, number, password, and money validators for better readability and consistency.
- Updated `pubspec.lock` to reflect new versions of dependencies.
- Added missing newline at the end of `main.dart` in `gmana_value_objects`.
- Enhanced `Money` class with better structure and immutability.

+++ packages/gmana_predicates/lib/predicates/numeric_predicates.dart
@@ -1,9 +1,10 @@
 /// Returns `true` if the numeric value of [str] is divisible by [n].
 ///
-/// Both [str] and [n] are parsed as numbers; returns `false` on parse failure.
+/// Both [str] and [n] are parsed as `double`; returns `false` on parse failure
+/// or when [n] is zero.
 bool isDivisibleBy(String str, String n) {
   try {
-    return double.parse(str) % int.parse(n) == 0;
+    return double.parse(str) % double.parse(n) == 0;
   } catch (_) {
     return false;
   }
